requirements.txt
pandas
Pillow
qrcode
cx_Freeze 